+++ lending-admin-client/src/app/app.module.ts
@@ -4,7 +4,6 @@ import {NgModule} from '@angular/core';
 import {AppRoutingModule} from './app-routing.module';
 import {AppComponent} from './app.component';
 import {NgbModule} from '@ng-bootstrap/ng-bootstrap';
-import {OverviewComponent} from './overview/overview.component';
 import {FormsModule} from '@angular/forms';
 import {ArchwizardModule} from 'ng2-archwizard';
 import {AddMultipleModule} from './add-multiple/add-multiple.module';
@@ -13,12 +12,11 @@ import {ActivateMultipleModule} from './activate-multiple/activate-multiple.modu
 import {AddSingleModule} from './add-single/add-single.module';
 import {ShowAllModule} from './show-all/show-all.module';
 import {ActivateSingleModule} from './activate-single/activate-single.module';
-import {ChartsModule} from 'ng2-charts';
+import {OverviewModule} from './overview/overview.module';
 
 @NgModule({
   declarations: [
-    AppComponent,
-    OverviewComponent
+    AppComponent
   ],
   imports: [
     BrowserModule,
@@ -31,7 +29,7 @@ import {ChartsModule} from 'ng2-charts';
     AddMultipleModule,
     ActivateSingleModule,
     ActivateMultipleModule,
-    ChartsModule
+    OverviewModule
   ],
   bootstrap: [AppComponent]
 })
