+++ lending-admin-client/src/app/core/core.module.ts
@@ -6,7 +6,7 @@ import {HttpClientModule} from '@angular/common/http';
 import {SnotifyService, ToastDefaults} from 'ng-snotify';
 import {IdentityCardService} from './identity-card.service';
 import {EnvelopeService} from './envelope.service';
-import {OverviewService} from './overview.service';
+import {OverviewService} from '../overview/overview.service';
 
 @NgModule({
   imports: [
@@ -19,8 +19,7 @@ import {OverviewService} from './overview.service';
     EntityService,
     GameService,
     IdentityCardService,
-    EnvelopeService,
-    OverviewService
+    EnvelopeService
   ]
 })
 export class CoreModule {
