--- lending-admin-client/src/app/overview/overview.component.css
@@ -1,5 +1,0 @@
-.overview-chart {
-  width: 100%;
-  max-height: 50%;
-  height: 300px;
-}
