+++ lending-admin-client/src/app/overview/overview/overview.component.css
@@ -1,0 +1,6 @@
+.overview-chart {
+  width: 100%;
+  min-height: 300px;
+  height: 33%;
+  max-height: 400px;
+}
